fix: theme manager
