+++ lib/prefs/appearance.js
@@ -8,7 +8,8 @@ import { gettext as _ } from "resource:///org/gnome/Shell/Extensions/js/extensio
 
 // Shared state
 import { ConfigManager } from "../shared/settings.js";
-import { ThemeManager } from "../shared/theme.js";
+
+import { PrefsThemeManager } from "./prefs-theme-manager.js";
 
 // Prefs UI
 import { ColorRow, PreferencesPage, ResetButton, SpinButtonRow, SwitchRow } from "./widgets.js";
@@ -36,7 +37,7 @@ export class AppearancePage extends PreferencesPage {
 
   constructor({ settings, dir }) {
     super({ title: _("Appearance"), icon_name: "brush-symbolic" });
-    this.themeMgr = new ThemeManager(this, { prefsMode: true });
+    this.themeMgr = new PrefsThemeManager(this);
     this.configMgr = new ConfigManager({ dir });
     this.add_group({
       title: _("Gaps"),
