+++ lib/shared/theme.js
@@ -20,31 +20,16 @@
 import Gio from "gi://Gio";
 import GLib from "gi://GLib";
 import GObject from "gi://GObject";
-import St from "gi://St";
 
 // Application imports
 import { stringify, parse } from "../css/index.js";
 import { production } from "./settings.js";
-import { Logger } from "./logger.js";
 
-// TODO: subclass this for the extension and prefs versions, rather than passing
-// in a flag
-export class ThemeManager extends GObject.Object {
-  static {
-    GObject.registerClass(this);
-  }
-
-  /** @type {string} */
-  extensionPath;
-
-  constructor({ configMgr, settings, dir, metadata, stylesheet }, options = { prefsMode: false }) {
+export class ThemeManagerBase extends GObject.Object {
+  constructor({ configMgr, settings }) {
     super();
     this.configMgr = configMgr;
     this.settings = settings;
-    this.metadata = metadata;
-    this.stylesheet = stylesheet;
-    if (options.prefsMode) this.extensionPath = `${dir.get_path()}`;
-    this.options = options;
     this._importCss();
     this.defaultPalette = this.getDefaultPalette();
 
@@ -59,10 +44,16 @@ export class ThemeManager extends GObject.Object {
     // TODO: should the patchCss() call be done here?
   }
 
+  /**
+   * @param {string} value
+   */
   addPx(value) {
     return `${value}px`;
   }
 
+  /**
+   * @param {string} value
+   */
   removePx(value) {
     return value.replace("px", "");
   }
@@ -80,6 +71,7 @@ export class ThemeManager extends GObject.Object {
   /**
    * The scheme name is in between the CSS selector name
    * E.g. window-tiled-color should return `tiled`
+   * @param {string} selector
    */
   getColorSchemeBySelector(selector) {
     if (!selector.includes("-")) return null;
@@ -89,6 +81,9 @@ export class ThemeManager extends GObject.Object {
     return scheme;
   }
 
+  /**
+   * @param {string} color
+   */
   getDefaults(color) {
     return {
       color: this.getCssProperty(`.${color}`, "color").value,
@@ -97,6 +92,9 @@ export class ThemeManager extends GObject.Object {
     };
   }
 
+  /**
+   * @param {any} selector
+   */
   getCssRule(selector) {
     if (this.cssAst) {
       const rules = this.cssAst.stylesheet.rules;
@@ -107,6 +105,10 @@ export class ThemeManager extends GObject.Object {
     return {};
   }
 
+  /**
+   * @param {string} selector
+   * @param {string} propertyName
+   */
   getCssProperty(selector, propertyName) {
     const cssRule = this.getCssRule(selector);
 
@@ -205,29 +207,7 @@ export class ThemeManager extends GObject.Object {
    * Credits: ExtensionSystem.js:_callExtensionEnable()
    */
   reloadStylesheet() {
-    if (this.options.prefsMode) {
-      this.settings.set_string("css-updated", Date.now().toString());
-    } else {
-      const uuid = this.metadata.uuid;
-      const stylesheetFile = this.configMgr.stylesheetFile;
-      const defaultStylesheetFile = this.configMgr.defaultStylesheetFile;
-      let theme = St.ThemeContext.get_for_stage(global.stage).get_theme();
-
-      try {
-        theme.unload_stylesheet(defaultStylesheetFile);
-        theme.unload_stylesheet(stylesheetFile);
-        if (production) {
-          theme.load_stylesheet(stylesheetFile);
-          this.stylesheet = stylesheetFile;
-        } else {
-          theme.load_stylesheet(defaultStylesheetFile);
-          this.stylesheet = defaultStylesheetFile;
-        }
-      } catch (e) {
-        Logger.error(`${uuid} - ${e}`);
-        return;
-      }
-    }
+    throw new Error("Must implement reloadStylesheet");
   }
 
   _needUpdate() {
