+++ tsconfig.json
@@ -3,15 +3,16 @@
     "lib/**/*",
     "extension.js",
     "prefs.js",
-    "@girs/adw-1",
-    "@girs/gio-2.0",
-    "@girs/gjs",
-    "@girs/gobject-2.0",
-    "@girs/gtk-4.0"
+    "./node_modules/@girs/adw-1",
+    "./node_modules/@girs/gio-2.0",
+    "./node_modules/@girs/gjs",
+    "./node_modules/@girs/gobject-2.0",
+    "./node_modules/@girs/gtk-4.0"
   ],
   "compilerOptions": {
     "allowJs": true,
     "checkJs": true,
+    "moduleResolution": "NodeNext",
     "lib": ["ES2022"],
     "target": "ES2022"
   }
